Move cmpexch_failure_memory_order from Generic.hpp to Common.hpp
